refactor: move get_index_names()
